+++ app/ui/cafe_car/ui/badge.rb
@@ -1,0 +1,10 @@
+module CafeCar
+  module UI
+    component :Badge, tag: :span do
+      flag :success
+      flag :warning
+      flag :danger
+      flag :info
+    end
+  end
+end